pyproject.toml
[tool.poetry]
name = "simple-messenger"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "simple_messenger"}]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
fastapi = "^0.89.1"
sqlalchemy = {extras = ["mypy"], version = "^2.0.1"}
python-jose = "^3.3.0"
python-socketio = {extras = ["client"], version = "^5.7.2"}
loguru = "^0.6.0"
passlib = "^1.7.4"
pydantic = {extras = ["dotenv"], version = "^1.10.4"}
asyncpg = "^0.27.0"
python-multipart = "^0.0.5"
alembic = "^1.9.2"
psycopg2-binary = "^2.9.5"


[tool.poetry.group.dev.dependencies]
mypy = "^0.991"
yapf = "^0.32.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
